requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
certifi==2018.11.29
cffi==1.11.5
cryptography==2.4.2
decorator==4.3.0
dj-database-url==0.5.0
Django==2.1.5
django-environ==0.4.5
django-heroku==0.3.1
factory-boy==2.11.1
Faker==1.0.1
flake8==3.6.0
future==0.17.1
gunicorn==19.9.0
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
mccabe==0.6.1
more-itertools==5.0.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.8.0
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
ptyprocess==0.6.0
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.3.1
pytest==4.1.0
python-dateutil==2.7.5
python-telegram-bot==11.1.0
pytz==2018.7
six==1.12.0
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
whitenoise==4.1.2
